Extract renameHashIdKey helper to transform.js
Move the renameHashIdKey utility function to a dedicated transform
module to keep cli.js within the 500-line file limit.
